Much improve, such lenses. Woo.
There has been a disturbing outbreak of lenses throughout the
code, and it is good. Some parts I am not happy about

a %~ (b %~ f)

seems like I "should" be able to use

a >>> b %~ f

but I think I must have fluffed up a lens somewhere, maybe that is
preventing me from doing this, or I've just not thought about this
enough. :)

More good things are coming and I'm pretty happy with how it is all
progressing.